separate parsing into its own source file
- helpful to handle complexity of multiple storage types
- start outlining parsing of different storage types
- goal of handling .lhe, .csv, .lhe.gz, and (maybe) tarballs (we'll see)
